Add github workflow files to publish ruby gem
This commit adds the github workflow files used to publish the gem to
the rubygems repo defined by the `GEM_HOST_API_KEY`.

It also includes a workflow to ensure the version is bumped for each
change, and so a `version.rb` file is added to track the version.

This also updates the gem's version number, combining the minor and
point release numbers as the minor number and using the point release
number to designate the custom SQM version (e.g. "2.2.3" becomes
"2.23.1"). This is done because the version-forget-me-not Github action
can only support semantic versioning with three numbers.