Add new types for schema validation, and update existing
This commit adds new SchemaType classes for use in structured data
validation.

Some types are updated to include new attributes.

The `contactPoints` property is superceded by `contactPoint` (per
schema.org`) and so the property is updated in `Organization`.

The `target` property of the `Action` type is updated to accept an
`EntryPoint` since that is what is being used in Viper.